feat: remove slashes from all pages
Previously, when directly visiting a page GitHub redirected users to a version
with a slash at the end. When clicking on a link React redirected to a version
without a slash at the end.

Because of this inconsitency, now all pages are rewritten from
for example /docs/index.html to /docs.html. GitHub Pages will make that
page accessible from /docs. The old place of the html will serve a
redirect.